Modularize short weierstrass curve templates
