+++ curves/src/bw6_761/tests.rs
@@ -28,7 +28,7 @@ use crate::{
         G2Projective,
         BW6_761,
     },
-    templates::short_weierstrass::tests::sw_tests,
+    templates::short_weierstrass_jacobian::tests::sw_tests,
     traits::{tests_curve::curve_tests, tests_group::group_test, AffineCurve, PairingEngine},
 };
 use snarkvm_fields::{
