+++ curves/src/templates/bls12/g1.rs
@@ -17,7 +17,7 @@
 use crate::{
     templates::{
         bls12::Bls12Parameters,
-        short_weierstrass::short_weierstrass_jacobian::{GroupAffine, GroupProjective},
+        short_weierstrass_jacobian::{GroupAffine, GroupProjective},
     },
     traits::pairing_engine::AffineCurve,
 };
