+++ curves/src/templates/short_weierstrass_jacobian/mod.rs
@@ -14,6 +14,11 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod short_weierstrass_jacobian;
-pub mod short_weierstrass_projective;
+pub mod affine;
+pub use affine::*;
+
+pub mod projective;
+pub use projective::*;
+
+#[cfg(test)]
 pub mod tests;
